requirements.txt
streamlit
langchain
langchain_openai
langchain-community
faiss-cpu
biopython
python-dotenv
pandas
plotly
python-docx==0.8.11
tenacity>=8.0.1
retry>=0.9.2